requirements.txt
alembic==1.13.0
aiohttp==3.9.2
asyncpg==0.29.0
fastapi==0.111.0
pydantic==1.10.17
pyjwt==2.8.0
python-dotenv==1.0.1
uvicorn==0.30.1
sqlalchemy==2.0.30
sqlalchemy[asyncio]==2.0.30
jinja2==3.1.4
bcrypt==5.0.0
